requirements.txt
beautifulsoup4==4.12.0
faiss_cpu==1.7.3
ffmpeg_python==0.2.0
load-dotenv==0.1.0
lxml==4.9.2
multilingual_clip==1.0.10
openai==0.27.2
pytest==7.2.2
scikit-learn==1.2.2
sseclient-py=1.7.2
temporalio==1.1.0
git+https://github.com/openai/CLIP.git@a9b1bf59 # main branch
